requirements.txt
black==25.9.0
fastapi==0.119.0
httpx==0.28.1
loguru==0.7.3
pydantic==2.12.3
pydantic_core==2.41.4
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
SQLAlchemy==2.0.44
uvicorn==0.38.0
